Implementing Color Emphasis (#85)
* First implementation of color emphasis

Still need to fix the color intensity; it's dimming and lighting up too much.

* Made the emphasis less intense and cleaned up some of the that code.

N.B.: The tinting is kind of a placeholder because we handle the colors entirely in RGB instead of the NES's composite encoding. This makes the colors less authentic, but makes it much easier to work with. It might need to be reworked depending on our needs.

* Some clean up before the PR.

* moved to_rgb() funcs outside Emulator (idk why I moved them there...)